requirements.txt
python-dotenv==1.1.0
langchain>=0.3.22,<0.4.0
langchain-openai>=0.3.11,<0.4.0
langchain-anthropic==0.3.10
langchain-ollama==0.3.1
langchain-google-genai==2.1.2
langgraph>=0.3.24,<0.4.0
tavily-python==0.5.4
rich==14.1.0
pyfiglet==1.0.3
qdrant-client
python-docx
python-pptx
mistralai
pymupdf
fastembed
rich==14.1.0
pyfiglet==1.0.3
pdfplumber
reportlab